remove all previous files & do new formatting
the purpose is for comparing my answer and feedback.
the problems start with number and dot. both files are the same
the answer and feedback start with as> and fb> each.
by calling git diff --no-index, you can check the differences of
your answers and feedbacks, with original problems are called once.