let version control system ignore .gitignore, also remove it from github
to tell git that some file already on git needs to be ignored,
use the command: git rm --cached the-file
if you need to ignore a directory that has many files in there, use:
git rm -r --cached . (in case you are in the directory)

after applying the proper command, add & commit as usual.
you can find it working.