refine words using Google search by reference. specification is below:
2. when search in Google, asset information isn't show the way as I
expected. But it seems to be o.k., just give it a try.

3. Judging from search, 'other trademarks are owned by their respective
owners' seems to be used as an idiom. I dig more in Quora and StackExchange,
but the answer uses words hard to understand. It seems other trademarks
of the document should be informed as property of respective companies
which own them so that break someone who thinks he/she can use the
(otherwise uninformed, so he/she thinks it can be used for their good) trademarks.