edit answers for m6 problems.
manual pdfs for topics in relation gives words in real use, though not always.

for the last problem, totally misunderstood what the problem means.
edit from the bottom.

for letter: googling, revising multiple time is my best at the moment, but is this
okay if I work for real job?